sketch of sink
